skills/ascii-art-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ascii-art-pipeline"
version = "0.1.0"
description = "High-fidelity ASCII rendering pipeline for images, video, and eikons"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Museah"}]
dependencies = [
  "Pillow>=10.0",
  "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
ascii-pipeline = "ascii_pipeline.cli:main"

[tool.setuptools.packages.find]
include = ["ascii_pipeline*"]
